Move `payment_history_change_notifier` to `controller`
